Another round of -Wconversion warning fixes, see [03cff7a777]

+++ win/tclWinPipe.c
@@ -3247,7 +3247,7 @@ TclpOpenTemporaryFile(
 	memcpy(namePtr, baseStr, length);
 	namePtr += length;
     }
-    counter = TclpGetClicks() % 65533;
+    counter = (int)(TclpGetClicks() % 65533);
     counter2 = 1024;			/* Only try this many times! Prevents
 					 * an infinite loop. */
